cathealth/requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
ultralytics
geopy
requests
python-dotenv